Move category bulk delete logic higher
